fix: renamed IEvent.ts to events.ts
